Added IS_score function and plot functions.
funcs.py:
Added a get_is score function as well as plotting functions for the loss, the FID scores and the IS scores as a function of epoch.

part_1.py:
Added lines to compute the FID and IS score at each epoch. And then lines at the end to plot them and get the scores for the fully trained model.